requirements.txt
streamlit
pandas
xlrd>=1.2.0
plotly==4.14.3
Pillow==8.1.2
protobuf~=3.19.0
openpyxl
numpy==1.22.4
python-dotenv
mysql-connector-python==8.0.28